backend/requirements/dev.txt
# ВНИМАНИЕ НЕОБХОДИМО ОБНОВИТЬ ДО ПОСЛЕДНИХ ВЕРСИЙ!

# backend/requirements/dev.txt
-r base.txt

# Разработка и отладка
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Автоперезагрузка и дебаг
watchfiles==0.21.0
debugpy==1.8.0

# Документация
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Тестирование (полное)
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
freezegun==1.2.2
httpx==0.25.1

# Инструменты БД
sqlalchemy-utils==0.41.1